First try to hook up twilio and react: sadness
Twilio has a bug where it thinks its running server side (my analysis)
and tries to use child_process, which is part of node's stdlib.
https://github.com/twilio/twilio-video.js/issues/28

I tried loading it in the html and then using it from the environment,
but something (webpack?) didn't like that it wasn't defined.

Here, I tried putting it in public and requireing it from there.
This works, but it takes several minutes to build the package
and its apparently not smart enough to cache and reuse the processed file,
so you have to pay that cost every time you build.

I looked through the build script that's invoked by `npm run build`,
and it doesn't take any arguments (at least nothing talks to process.argv)
so I'm committing the progress, but going to have to find something better
b/c that cost will kill my chances of succeeding.