Tokens are timestamps in base36 instead of digests in base64
We can't use base64 because it can output slashes, which mess up Sinatra's routing.
(a regex can probably get around this, but it would be confusing to anyone reading the path, too)

Was going to switch to bubble_babble, but it was so long that it was annoying, eg:
Digest::MD5.bubblebabble(str) # => "xusak-pyged-nytyh-divoh-pahet-rynek-kocip-micic-huxux"

Was going to only use the first two bubbles, but IDK how that affects the uniqueness probability,
and the timestamp is enough to be unique for now, anyway.

Just representing it with Base36 is simpler, probably faster, and I understand what it will emit.
Here is an example output:

time.strftime("%Y%m%d%H%M%9N").to_i.to_s(36) # => "16jr6dh3918qy0"